Add option to show off-hand hit markers on main
Make it possible to show a marker on the main swing bar when the off
hand is expected to hit. A new configuration option to "Show Off-Hand
on Main" is added to toggle this behavior.

This allows a heroic strike queue enthusiast to only display the main
swing bar and from that decide whether to queue heroic strike or not.